Moved more tests out of EntityTest
